# Improve materialize_batch_and_maybe_update_state