--- a/server/Cargo.toml
+++ b/server/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "server"
-version = "0.4.84"
+version = "0.4.85"
 edition = "2021"
 build = "src/build.rs"