Refactor materialize_batch_and_maybe_update_state to use drain for processing message batches, eliminating unnecessary copies of self.messages and enhancing performance.

Iggy bench results for **messages_required_to_save = 1**(via `cargo run --release --bin iggy-bench -- send tcp`)

new function:
```
2024-12-07T15:31:12.909732Z  INFO iggy_bench::benchmark_runner: Results: Total throughput: 971.09 MB/s, 948330 messages/s, average throughput: 97.11 MB/s, average p50 latency: 9.00 ms, average p90 latency: 17.96 ms, average p95 latency: 22.11 ms, average p99 latency: 31.07 ms, average p999 latency: 74.38 ms, average latency: 10.55 ms, median latency: 9.00 ms
```

old:
```
2024-12-07T15:24:44.084381Z  INFO iggy_bench::benchmark_runner: Results: Total throughput: 129.75 MB/s, 126708 messages/s, average throughput: 12.97 MB/s, average p50 latency: 35.13 ms, average p90 latency: 104.77 ms, average p95 latency: 283.10 ms, average p99 latency: 1101.02 ms, average p999 latency: 2727.17 ms, average latency: 78.92 ms, median latency: 35.13 ms
```

**messages_required_to_save=5000:**

new
```
5000 2024-12-08T05:26:41.258931Z  INFO iggy_bench::benchmark_runner: Results: Total throughput: 1678.50 MB/s, 1639159 messages/s, average throughput: 167.85 MB/s, average p50 latency: 4.16 ms, average p90 latency: 11.71 ms, average p95 latency: 15.38 ms, average p99 latency: 25.87 ms, average p999 latency: 88.84 ms, average latency: 6.10 ms, median latency: 4.16 ms
```

old:
```
2024-12-08T05:34:17.895922Z  INFO iggy_bench::benchmark_runner: Results: Total throughput: 1591.14 MB/s, 1553850 messages/s, average throughput: 159.11 MB/s, average p50 latency: 4.47 ms, average p90 latency: 13.01 ms, average p95 latency: 17.10 ms, average p99 latency: 28.18 ms, average p999 latency: 44.04 ms, average latency: 6.44 ms, median latency: 4.47 ms
```

The new function significantly improves performance for low values of messages_required_to_save, especially in throughput and latency. For higher values, the performance gap narrows