@haze518 2024-12-08
Iggy bench results for messages_required_to_save = 1(via `cargo run --release --bin iggy-bench -- send tcp`)

new function:
```
2024-12-07T15:31:12.909732Z  INFO iggy_bench::benchmark_runner: Results: Total throughput: 971.09 MB/s, 948330 messages/s, average throughput: 97.11 MB/s, average p50 latency: 9.00 ms, average p90 latency: 17.96 ms, average p95 latency: 22.11 ms, average p99 latency: 31.07 ms, average p999 latency: 74.38 ms, average latency: 10.55 ms, median latency: 9.00 ms
```

old:
```
2024-12-07T15:24:44.084381Z  INFO iggy_bench::benchmark_runner: Results: Total throughput: 129.75 MB/s, 126708 messages/s, average throughput: 12.97 MB/s, average p50 latency: 35.13 ms, average p90 latency: 104.77 ms, average p95 latency: 283.10 ms, average p99 latency: 1101.02 ms, average p999 latency: 2727.17 ms, average latency: 78.92 ms, median latency: 35.13 ms
```